+++ OpenHome/Net/Bindings/Java/CpDeviceListUpnpAll.c
@@ -38,7 +38,7 @@ JNIEXPORT jobject JNICALL Java_org_openhome_net_controlpoint_CpDeviceListUpnpAll
 	}
 	
 	devList = CpDeviceListCreateUpnpAll(callbackAdded, ref, callbackRemoved, ref);
-	devListInit = (*aEnv)->NewObject(aEnv, statusClass, cid, aObject, (jlong)devList, (jlong)ref);
+	devListInit = (*aEnv)->NewObject(aEnv, statusClass, cid, aObject, (jlong)(size_t)devList, (jlong)(size_t)ref);
 	
 	return devListInit;
 }
