+++ OpenHome/Net/Bindings/Java/ParameterUint.c
@@ -20,7 +20,7 @@ JNIEXPORT jlong JNICALL Java_org_openhome_net_core_ParameterUint_ServiceParamete
 	
 	(*aEnv)->ReleaseStringUTFChars(aEnv, aName, name);
 	
-	return (jlong) param;
+	return (jlong) (size_t)param;
 }
 
 #ifdef __cplusplus
