+++ OpenHome/Net/Bindings/Java/PropertyBool.c
@@ -1,7 +1,7 @@
 #include <jni.h>
 #include "PropertyCallback.h"
 #include "PropertyBool.h"
-#include "OpenHome/Net/C/Ohnet.h"
+#include "OpenHome/Net/C/OhNet.h"
 #include "OpenHome/Net/C/Service.h"
 
 #ifdef __cplusplus
@@ -38,7 +38,7 @@ JNIEXPORT jobject JNICALL Java_org_openhome_net_core_PropertyBool_ServicePropert
 	}
 	
 	property = ServicePropertyCreateBoolCp(name, callback, ref);
-	propertyInit = (*aEnv)->NewObject(aEnv, statusClass, cid, aObject, (jlong)property, (jlong)ref);
+	propertyInit = (*aEnv)->NewObject(aEnv, statusClass, cid, aObject, (jlong)(size_t)property, (jlong)(size_t)ref);
 	
 	(*aEnv)->ReleaseStringUTFChars(aEnv, aName, name);
 	
@@ -58,7 +58,7 @@ JNIEXPORT jlong JNICALL Java_org_openhome_net_core_PropertyBool_ServicePropertyC
 	aEnv = aEnv;
 	aClass = aClass;
 	
-	return (jlong) property;
+	return (jlong) (size_t)property;
 }
 
 /*
